Adds support for Pokestop spinning.
Use --spin-below-level <lvl> and --max-spins <num> to configure for which accounts Pokestops should be spinned and how many at most. Default vaules are spin-below-level: 2 and max-spins: 1, so that there will be ONE Pokestop spin if an account is level 1. Of course your scan location must be in range of one or multiple Pokestops.